move config file to config dir
