+++ lib/controller.js
@@ -4,9 +4,9 @@ const dotenv = require('dotenv');
 const _ = require('lodash');
 const { error } = dotenv.config({ silent: true });
 const axios = require('axios');
-const config = require('./config');
+const apiConfig = require('./config/api');
 
-const apiUrl = config.GITHUB_API_URL;
+const apiUrl = apiConfig.GITHUB_API_URL;
 
 // load environment file
 if (error) {
